backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator==2.1.0

# AI / LLM
litellm>=1.17.9

# AWS S3
boto3==1.34.14

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
Pillow>=11.0.0
